Improvements and drag&drop for files

+++ src/components/Explorer/components/directory-component.tsx
@@ -1,13 +1,14 @@
 import React from "react";
 import {VscChevronDown, VscChevronRight} from "react-icons/vsc";
 
+import {FileBasic} from "@utils/file-system/basic";
 import {Directory} from "@utils/file-system/directory";
 import {File} from "@utils/file-system/file";
 
 import {ContextMenu} from "@components/ContextMenu";
 
 import {useAppDispatch, useAppSelector} from "@redux/hooks";
-import {renameDirectory, setFileTreeStates} from "@redux/reducers/files";
+import {renameDirectory, renameFile, setFileTreeStates} from "@redux/reducers/files";
 import {addNotification} from "@redux/reducers/notifications";
 
 import {NotificationType} from "@shared-types/notifications";
@@ -17,9 +18,12 @@ import {v4} from "uuid";
 import {FileComponent} from "./file-component";
 import {NewItem, NewItemType} from "./new-item";
 
+import {ContextMenuTemplate} from "../../ContextMenu/context-menu";
+
 export type DirectoryComponentProps = {
     level: number;
     directory: Directory;
+    rootDirectory?: boolean;
 };
 
 export const DirectoryComponent: React.VFC<DirectoryComponentProps> = props => {
@@ -84,7 +88,7 @@ export const DirectoryComponent: React.VFC<DirectoryComponentProps> = props => {
     }, [props.directory, dispatch]);
 
     const contextMenuTemplate = React.useMemo(() => {
-        return [
+        const template: ContextMenuTemplate = [
             {
                 label: "New File...",
                 click: () => {
@@ -103,23 +107,28 @@ export const DirectoryComponent: React.VFC<DirectoryComponentProps> = props => {
                     setCreatingNewDir(true);
                 },
             },
-            {
-                divider: true,
-            },
-            {
-                label: "Rename...",
-                click: () => {
-                    setEditMode(true);
-                },
-            },
-            {
-                label: "Delete",
-                click: () => {
-                    handleDelete();
-                },
-            },
         ];
-    }, [handleDelete, handleDirStateChange, expanded]);
+        if (!props.rootDirectory) {
+            template.push(
+                {
+                    divider: true,
+                },
+                {
+                    label: "Rename...",
+                    click: () => {
+                        setEditMode(true);
+                    },
+                },
+                {
+                    label: "Delete",
+                    click: () => {
+                        handleDelete();
+                    },
+                }
+            );
+        }
+        return template;
+    }, [handleDelete, handleDirStateChange, expanded, props.rootDirectory]);
 
     const handleSubmit = React.useCallback(
         (name: string) => {
@@ -161,9 +170,11 @@ export const DirectoryComponent: React.VFC<DirectoryComponentProps> = props => {
         setDirName(props.directory.baseName());
     }, [props.directory]);
 
-    if (deleted) {
-        return null;
-    }
+    const handleDragStart = (e: React.DragEvent<HTMLDivElement>) => {
+        e.stopPropagation();
+        e.dataTransfer.effectAllowed = "move";
+        e.dataTransfer.setData("text/plain", props.directory.relativePath());
+    };
 
     const handleDragOver = (e: React.DragEvent<HTMLDivElement>) => {
         e.preventDefault();
@@ -177,52 +188,83 @@ export const DirectoryComponent: React.VFC<DirectoryComponentProps> = props => {
         setDragOver(false);
     };
 
-    const handleDrop = (e: React.DragEvent<HTMLDivElement>) => {
-        e.preventDefault();
-        e.stopPropagation();
-        setDragOver(false);
-    };
+    const handleDrop = React.useCallback(
+        (e: React.DragEvent<HTMLDivElement>) => {
+            e.preventDefault();
+            e.stopPropagation();
+            setDragOver(false);
+            const droppedAsset = new FileBasic(
+                e.dataTransfer.getData("text/plain"),
+                props.directory.workingDirectory()
+            );
+            const oldPath = droppedAsset.absolutePath();
+            if (droppedAsset.moveToDir(props.directory.absolutePath())) {
+                dispatch(renameFile({oldFilePath: oldPath, newFilePath: droppedAsset.absolutePath()}));
+                dispatch(
+                    addNotification({
+                        type: NotificationType.SUCCESS,
+                        message: `'${droppedAsset.relativePath()}' successfully moved to '${props.directory.relativePath()}'.`,
+                    })
+                );
+            } else {
+                dispatch(
+                    addNotification({
+                        type: NotificationType.ERROR,
+                        message: `'${droppedAsset.relativePath()}' could not be moved to '${props.directory.relativePath()}'.`,
+                    })
+                );
+            }
+        },
+        [props.directory, dispatch]
+    );
+
+    if (deleted) {
+        return null;
+    }
 
     /* eslint-disable jsx-a11y/no-autofocus */
     return (
         <div
-            className={`Directory${dragOver ? " DirectoryDragOver" : ""}`}
+            className={`Directory${dragOver ? " DirectoryDragOver" : ""}${props.rootDirectory ? " RootDirectory" : ""}`}
             ref={ref}
+            onDragStart={handleDragStart}
             onDragOver={handleDragOver}
             onDragLeave={handleDragLeave}
             onDrop={handleDrop}
         >
             <ContextMenu parent={ref.current} template={contextMenuTemplate} />
-            <a
-                className="ExplorerItem"
-                href="#"
-                onClick={e => handleDirStateChange(e)}
-                title={props.directory.relativePath()}
-            >
-                {props.level > 1 &&
-                    Array(props.level - 1)
-                        .fill(0)
-                        .map(_ => <div className="ExplorerPath" key={`${props.directory.baseName()}-${v4()}}`} />)}
-                <div className="ExplorerItemIcon">
-                    {expanded ? <VscChevronDown fontSize="small" /> : <VscChevronRight fontSize="small" />}
-                </div>
-                {editMode ? (
-                    <>
-                        <div className="Overflow" onClick={() => onClose()} />
-                        <input
-                            className="ExplorerItemInput"
-                            autoFocus
-                            type="text"
-                            defaultValue={dirName}
-                            onKeyDown={handleKeyDown}
-                        />
-                    </>
-                ) : (
-                    <div className="ExplorerItemText">{dirName}</div>
-                )}
-            </a>
+            {!props.rootDirectory && (
+                <a
+                    className="ExplorerItem"
+                    href="#"
+                    onClick={e => handleDirStateChange(e)}
+                    title={props.directory.relativePath()}
+                >
+                    {props.level > 1 &&
+                        Array(props.level - 1)
+                            .fill(0)
+                            .map(_ => <div className="ExplorerPath" key={`${props.directory.baseName()}-${v4()}}`} />)}
+                    <div className="ExplorerItemIcon">
+                        {expanded ? <VscChevronDown fontSize="small" /> : <VscChevronRight fontSize="small" />}
+                    </div>
+                    {editMode ? (
+                        <>
+                            <div className="Overflow" onClick={() => onClose()} />
+                            <input
+                                className="ExplorerItemInput"
+                                autoFocus
+                                type="text"
+                                defaultValue={dirName}
+                                onKeyDown={handleKeyDown}
+                            />
+                        </>
+                    ) : (
+                        <div className="ExplorerItemText">{dirName}</div>
+                    )}
+                </a>
+            )}
             <div className="DirectoryContent">
-                {expanded && (
+                {(expanded || props.rootDirectory) && (
                     <>
                         {props.directory && creatingNewDir && (
                             <NewItem
